+++ lib/src/filters/wave_shaper_filter.dart
@@ -8,7 +8,6 @@ enum WaveShaperEnum {
   wet,
   amount;
 
-  /// use iterables?
   final List<double> _mins = const [0, -1];
   final List<double> _maxs = const [1, 1];
   final List<double> _defs = const [1, 0];
@@ -24,20 +23,15 @@ enum WaveShaperEnum {
       };
 }
 
-abstract class WaveShaperInternal {
-  const WaveShaperInternal(SoundHash? soundHash) : _soundHash = soundHash;
+abstract class _WaveShaperInternal extends FilterBase {
+  const _WaveShaperInternal(SoundHash? soundHash)
+      : super(FilterType.waveShaperFilter, soundHash);
 
-  final SoundHash? _soundHash;
-  FilterType get filterType => FilterType.waveShaperFilter;
   WaveShaperEnum get queryWet => WaveShaperEnum.wet;
   WaveShaperEnum get queryAmount => WaveShaperEnum.amount;
-
-  void activate() => filterType.activate(_soundHash);
-
-  void deactivate() => filterType.deactivate(_soundHash);
 }
 
-class WaveShaperSingle extends WaveShaperInternal {
+class WaveShaperSingle extends _WaveShaperInternal {
   WaveShaperSingle(super.soundHash);
 
   FilterParam wet({SoundHandle? soundHandle}) => FilterParam(
@@ -57,7 +51,7 @@ class WaveShaperSingle extends WaveShaperInternal {
       );
 }
 
-class WaveShaperGlobal extends WaveShaperInternal {
+class WaveShaperGlobal extends _WaveShaperInternal {
   const WaveShaperGlobal() : super(null);
 
   FilterParam get wet => FilterParam(